Refactored Lottie animations
